adding authorization source code

+++ app/Http/Controllers/ListingController.php
@@ -1,15 +1,28 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\Http\Middleware\NotSuspended;
 use App\Models\Listing;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Inertia\Inertia;
 
-class ListingController extends Controller
+class ListingController extends Controller implements HasMiddleware
 {
+
+    public static function middleware() {
+        return [
+            new Middleware(
+                ['auth', 'verified', NotSuspended::class],
+                except: ['index', 'show']
+            )
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -36,6 +49,7 @@ class ListingController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Listing::class);
         return Inertia::render('Listing/Create');
     }
 
@@ -75,9 +89,12 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+        Gate::authorize('view', $listing);
+
         return Inertia::render('Listing/Show',[
             'listing'=> $listing,
-            'user'=> $listing->user->only(['name', 'id'])
+            'user'=> $listing->user->only(['name', 'id']),
+            'canModify' => Auth::user() ? Auth::user()->can('modify', $listing) : false
         ]);
     }
 
@@ -86,6 +103,8 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
+        Gate::authorize('modify', $listing);
+
         return Inertia::render('Listing/Edit', [
             'listing'=>$listing
         ]);
@@ -96,6 +115,8 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        Gate::authorize('modify', $listing); 
+
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
             'desc' => ['required'],
@@ -126,6 +147,8 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('modify', $listing);
+
         if ($listing->image) {
             Storage::disk('public')->delete($listing->image);
         }
